Introduce initial ECS component
This module is intended to be used with an ALB module still in
development.

This module requires a lot more consumer level code than the previous
container.py, however it uses an atomic component based approach instead
of a monolithic Component.  This means that with the AlbComponent, we
can compose new service types with this EcsCompoment.

The service repository will still have minimal infrastructure code, if
we introduce a concept of "service_types" in this repository, that
layers on top of this EcsComponent.  Things to come in the future!