BEL-3138 Fix autoscaling issues on creation.
Deployment does not work the first time because it depends on the fargate
service already being available to set an autoscaling target. This was not the case.

Folks were turning autoscaling off, and then turning it back on again in order to make this work.

Specifying the dependency order fixes this problem.