BEL-4567 Allow canvas components to work
Because we were constructing the name and using the namespace we ended up doubling up the names for canvas components. Instead of using the namespace to guess what the cluster is called, we change to sending the actual cluster to the ContainerComponent.
We also fixed the comparison_operator so that we will alarm when the threshold is met because we intended to alarm at 100 not 101

Co-authored-by: mark ng <mark.ng@strongmind.com>